pyproject.toml
[project]
name = "eve"
version = "0.3.6"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "qwen-asr",
  "rich",
  "silero-vad",
  "soundfile",
  "sounddevice",
  "flet>=0.82.2",
  "pystray>=0.19.5",
  "pillow>=12.1.0",
  "platformdirs>=4.5.1",
  "pyobjc-framework-avfoundation>=12.1 ; sys_platform == 'darwin'",
]

[project.scripts]
eve = "eve.cli:main"
eve-desktop = "eve.desktop_app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true
